Add files via upload
In this project, we shall explore the posts in the hacker news website. Both the Ask Hacker News and Show hacker news. The ask Hacker is used by users to ask the community some questions. The show hacker news is used by users to show the community some stuffs they did.